ng-system-common-app/projects/ng-system-common/src/lib/collections/Enumerable.ts
import { IEnumerable } from './IEnumerable';
import { EnumerableIterator } from './EnumerableIterator';

export class Enumerable<T> implements IEnumerable<T> {
    private items: T[] = [];

    /** Get counts of collection */
    get Count(): number {
        return this.Items.length;
    }
    /** Get all items */
    get Items(): T[] {
        return this.items;
    }
    /** Set items value */
    set Items(value: T[]) {
        this.items = value;
    }

    [Symbol.iterator]() {
        return new EnumerableIterator(this);
    }

    static From<T>(array: Array<T>): IEnumerable<T> {
        let list = new Enumerable<T>();
        if (array) {
            list.Items = array;
        }
        return list;
    }
}

